feat: Complete Owner-requested updates
- Fixed branch logic for Walk Ins, Renewals, Freezes, and Monthly Clients
- Added monthly client attendance logs
- Introduced a button to sum Grand Petty and Grand Expenses
- Enhanced Staff Scheduling:
    - Morning Shift: 5:30 AM to 2:30 PM
    - Mid Shift: 10:00 AM to 7:00 PM
    - Evening Shift: 3:00 PM to 12:00 MN
    - Dynamic Shift adjustments
- Implemented Payroll Column Filters (by Date Range and Branch)
- Enabled Daily Cash Flow edit functionality
- Improved Coach Booking Logic:
    - Allow 1 to 3 members per booking hour
    - Made coach schedule dynamic
    - Added notifications to coaches with booking details

+++ database/migrations/2025_01_10_013000_create_invoices_table.php
@@ -16,8 +16,12 @@ return new class extends Migration
             $table->unsignedBigInteger('BranchID')->nullable(); 
             $table->foreign('BranchID')->references('BranchID')->on('branches');
             $table->unsignedBigInteger('MemberID')->nullable();
-            $table->unsignedBigInteger('MonthlyClientID')->nullable();
-            $table->foreign('MonthlyClientID')->references('MonthlyClientID')->on('monthly_clients')->onDelete('cascade'); // or set null
+            $table->unsignedInteger('MonthlyClientID')->nullable();
+            $table
+              ->foreign('MonthlyClientID')
+              ->references('MonthlyClientID')
+              ->on('monthly_clients')
+              ->onDelete('cascade');            
             $table->unsignedBigInteger('PromotionID')->nullable();
             $table->dateTime('InvoiceDate')->nullable();
             $table->date('DueDate')->nullable();
